+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Functions used in global scope (decorator, default args) are sorted correctly.
+
 
 ## [0.1.3] - 2024-08-12